feat/#socketio: Add something new

+++ modules/restaurant/biz/delete_restaurant.go
@@ -15,11 +15,12 @@ type DeleteRestaurantStore interface {
 }
 
 type deleteRestaurantBiz struct {
-	store DeleteRestaurantStore
+	store     DeleteRestaurantStore
+	requester common.Requester
 }
 
-func NewDeleteRestaurantBiz(store DeleteRestaurantStore) *deleteRestaurantBiz {
-	return &deleteRestaurantBiz{store: store}
+func NewDeleteRestaurantBiz(store DeleteRestaurantStore, requester common.Requester) *deleteRestaurantBiz {
+	return &deleteRestaurantBiz{store: store, requester: requester}
 }
 
 func (biz *deleteRestaurantBiz) DeleteRestaurant(ctx context.Context, id uint32) error {
@@ -34,6 +35,10 @@ func (biz *deleteRestaurantBiz) DeleteRestaurant(ctx context.Context, id uint32)
 
 	if oldData.Status == 0 {
 		return errors.New("restaurant has been deleted")
+	}
+
+	if oldData.UserId != biz.requester.GetUserId() {
+		return common.ErrNotPermission(nil)
 	}
 
 	if err := biz.store.Delete(ctx, id); err != nil {
